+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to the External Files VS Code Extension will be documented i
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
-## 1.1.0 - 2025-08-??
+## 1.1.0 - 2025-08-08
 
 ### Added